+++ README.md
@@ -4,13 +4,26 @@
 **Windows:**
 Execute the following commands, or simply run the included run.bat file
 ```batch
-pip insall -r requirements.txt
-start cmd /k python webApp.py
-start cmd /k python -m http.server --directory WebApp
+pip install -r WebApp/requirements.txt
+start cmd /k python WebApp/webApp.py
 ```
+
+**Linux:**
+Execute the following commands, or simply run the included run.sh file
+```bash
+pip install -r WebApp/requirements.txt
+python3 WebApp/webApp.py
+```
+
+Remember to put the apiToken.txt file in /WebApp with the API tokens in this order:
+```txt
+<västtrafik_token>
+<styr&ställ_token>
+```
+
+By default the site is opened on localhost:5000
 
 ### Dependencies
 * Python 3
+* pip
 * Flask
-* Flask-cors
-* Any webserver
