+++ WebApp/run.bat
@@ -1,5 +1,5 @@
-pip insall -r requirements.txt
+pip install -r requirements.txt
 start cmd /k python webApp.py
 start cmd /k python -m http.server --directory WebApp
-explorer http://localhost:8000
+explorer http://localhost:5000
 ```
