+++ WebApp/run.sh
@@ -1,0 +1,2 @@
+pip install -r requirements.txt
+python3 webApp.py
